refactor(npm): clean up and prepare modules for usage in node
