+++ npm/packages/@socketsupply/socket-linux-arm64/package.json
@@ -4,16 +4,12 @@
   "description": "A Cross-Platform, Native Runtime for Desktop and Mobile Apps — Create apps using HTML, CSS, and JavaScript. Written from the ground up to be small and maintainable.",
   "type": "module",
   "main": "src/index.js",
-  "preferGlobal": true,
   "os": [
     "linux"
   ],
   "cpu": [
     "arm64"
   ],
-  "bin": {
-    "ssc": "bin/ssc-platform.js"
-  },
   "scripts": {
     "test": ":",
     "install": "bin/verify-platform.js linux arm64"
