+++ npm/packages/@socketsupply/socket-win32-x64/package.json
@@ -4,16 +4,12 @@
   "description": "A Cross-Platform, Native Runtime for Desktop and Mobile Apps — Create apps using HTML, CSS, and JavaScript. Written from the ground up to be small and maintainable.",
   "type": "module",
   "main": "src/index.js",
-  "preferGlobal": true,
   "os": [
     "win32"
   ],
   "cpu": [
     "x64"
   ],
-  "bin": {
-    "ssc": "bin/ssc-platform.js"
-  },
   "scripts": {
     "test": ":",
     "install": "cd bin && node verify-platform.js win32 x64"
