--- npm/packages/@socketsupply/socket/index.js
@@ -1,8 +1,0 @@
-import os from 'node:os'
-
-export async function load () {
-  const platform = os.platform()
-  const arch = os.arch()
-  const info = await import(`@socketsupply/socket-${platform}-${arch}`)
-  return info?.default ?? info ?? null
-}
